refactor: move resultProxy from bridge.go to result.go
Move resultProxy type and its methods alongside the Result
type and other result-related code. No functional changes.
