+++ solver/llbsolver/bridge.go
@@ -18,15 +18,11 @@ import (
 	"github.com/moby/buildkit/identity"
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/solver"
-	"github.com/moby/buildkit/solver/errdefs"
-	llberrdefs "github.com/moby/buildkit/solver/llbsolver/errdefs"
-	"github.com/moby/buildkit/solver/llbsolver/provenance"
 	"github.com/moby/buildkit/solver/pb"
 	"github.com/moby/buildkit/sourcepolicy"
 	spb "github.com/moby/buildkit/sourcepolicy/pb"
 	"github.com/moby/buildkit/util/bklog"
 	"github.com/moby/buildkit/util/entitlements"
-	"github.com/moby/buildkit/util/flightcontrol"
 	"github.com/moby/buildkit/util/progress"
 	"github.com/moby/buildkit/worker"
 	digest "github.com/opencontainers/go-digest"
@@ -214,145 +210,6 @@ func (b *llbBridge) loadExecutor() error {
 		b.executor = w.Executor()
 	})
 	return b.executorErr
-}
-
-type resultProxy struct {
-	id         string
-	b          *provenanceBridge
-	req        frontend.SolveRequest
-	g          flightcontrol.Group[solver.CachedResult]
-	mu         sync.Mutex
-	released   bool
-	v          solver.CachedResult
-	err        error
-	errResults []solver.Result
-	provenance *provenance.Capture
-}
-
-func newResultProxy(b *provenanceBridge, req frontend.SolveRequest) *resultProxy {
-	return &resultProxy{req: req, b: b, id: identity.NewID()}
-}
-
-func (rp *resultProxy) ID() string {
-	return rp.id
-}
-
-func (rp *resultProxy) Definition() *pb.Definition {
-	return rp.req.Definition
-}
-
-func (rp *resultProxy) Provenance() any {
-	if rp.provenance == nil {
-		return nil
-	}
-	return rp.provenance
-}
-
-func (rp *resultProxy) Release(ctx context.Context) (err error) {
-	rp.mu.Lock()
-	defer rp.mu.Unlock()
-	for _, res := range rp.errResults {
-		rerr := res.Release(ctx)
-		if rerr != nil {
-			err = rerr
-		}
-	}
-	if rp.v != nil {
-		if rp.released {
-			bklog.G(ctx).Warnf("release of already released result")
-		}
-		rerr := rp.v.Release(ctx)
-		if err != nil {
-			return rerr
-		}
-	}
-	rp.released = true
-	return
-}
-
-func (rp *resultProxy) wrapError(err error) error {
-	if err == nil {
-		return nil
-	}
-	var ve *errdefs.VertexError
-	if errors.As(err, &ve) {
-		if rp.req.Definition.Source != nil {
-			locs, ok := rp.req.Definition.Source.Locations[ve.Digest]
-			if ok {
-				for _, loc := range locs.Locations {
-					err = errdefs.WithSource(err, &errdefs.Source{
-						Info:   rp.req.Definition.Source.Infos[loc.SourceIndex],
-						Ranges: loc.Ranges,
-					})
-				}
-			}
-		}
-	}
-	return err
-}
-
-func (rp *resultProxy) loadResult(ctx context.Context) (solver.CachedResultWithProvenance, error) {
-	res, err := rp.b.loadResult(ctx, rp.req.Definition, rp.req.CacheImports, rp.req.SourcePolicies)
-	var ee *llberrdefs.ExecError
-	if errors.As(err, &ee) {
-		ee.EachRef(func(res solver.Result) error {
-			rp.errResults = append(rp.errResults, res)
-			return nil
-		})
-		// acquire ownership so ExecError finalizer doesn't attempt to release as well
-		ee.OwnerBorrowed = true
-	}
-	return res, err
-}
-
-func (rp *resultProxy) Result(ctx context.Context) (res solver.CachedResult, err error) {
-	defer func() {
-		err = rp.wrapError(err)
-	}()
-	return rp.g.Do(ctx, "result", func(ctx context.Context) (solver.CachedResult, error) {
-		rp.mu.Lock()
-		if rp.released {
-			rp.mu.Unlock()
-			return nil, errors.Errorf("accessing released result")
-		}
-		if rp.v != nil || rp.err != nil {
-			rp.mu.Unlock()
-			return rp.v, rp.err
-		}
-		rp.mu.Unlock()
-		v, err := rp.loadResult(ctx)
-		if err != nil {
-			select {
-			case <-ctx.Done():
-				if errdefs.IsCanceled(ctx, err) {
-					return v, err
-				}
-			default:
-			}
-		}
-		rp.mu.Lock()
-		if rp.released {
-			if v != nil {
-				v.Release(context.TODO())
-			}
-			rp.mu.Unlock()
-			return nil, errors.Errorf("evaluating released result")
-		}
-		if err == nil {
-			var capture *provenance.Capture
-			capture, err = captureProvenance(ctx, v)
-			if err != nil {
-				err = errors.Errorf("failed to capture provenance: %v", err)
-				v.Release(context.TODO())
-				v = nil
-			}
-			rp.provenance = capture
-		}
-		rp.v = v
-		rp.err = err
-		rp.mu.Unlock()
-		return v, err
-	})
 }
 
 func (b *llbBridge) ResolveSourceMetadata(ctx context.Context, op *pb.SourceOp, opt sourceresolver.Opt) (resp *sourceresolver.MetaResponse, err error) {
